Selection filter - refactor (move selection filter out of select.h/cpp)
