+++ src/engraving/dom/select.h
@@ -28,6 +28,8 @@
 #include "pitchspelling.h"
 #include "types.h"
 
+#include "selectionfilter.h"
+
 namespace mu::engraving {
 class Score;
 class Page;
@@ -83,64 +85,6 @@ enum class SelState : char {
     LIST,     // disjoint selection
     RANGE,    // adjacent selection, a range in one or more staves
     // is selected
-};
-
-//---------------------------------------------------------
-//   SelectionFilterType
-//---------------------------------------------------------
-
-static constexpr size_t NUMBER_OF_SELECTION_FILTER_TYPES = 23;
-
-enum class SelectionFilterType : unsigned int {
-    NONE                    = 0,
-    FIRST_VOICE             = 1 << 0,
-    SECOND_VOICE            = 1 << 1,
-    THIRD_VOICE             = 1 << 2,
-    FOURTH_VOICE            = 1 << 3,
-    DYNAMIC                 = 1 << 4,
-    HAIRPIN                 = 1 << 5,
-    FINGERING               = 1 << 6,
-    LYRICS                  = 1 << 7,
-    CHORD_SYMBOL            = 1 << 8,
-    OTHER_TEXT              = 1 << 9,
-    ARTICULATION            = 1 << 10,
-    ORNAMENT                = 1 << 11,
-    SLUR                    = 1 << 12,
-    FIGURED_BASS            = 1 << 13,
-    OTTAVA                  = 1 << 14,
-    PEDAL_LINE              = 1 << 15,
-    OTHER_LINE              = 1 << 16,
-    ARPEGGIO                = 1 << 17,
-    GLISSANDO               = 1 << 18,
-    FRET_DIAGRAM            = 1 << 19,
-    BREATH                  = 1 << 20,
-    TREMOLO                 = 1 << 21,
-    GRACE_NOTE              = 1 << 22,
-    ALL                     = ~(~0u << NUMBER_OF_SELECTION_FILTER_TYPES)
-};
-
-//---------------------------------------------------------
-//   SelectionFilter
-//---------------------------------------------------------
-
-class SelectionFilter
-{
-public:
-    SelectionFilter() = default;
-    SelectionFilter(SelectionFilterType type);
-
-    inline bool operator==(const SelectionFilter& f) const { return m_filteredTypes == f.m_filteredTypes; }
-    inline bool operator!=(const SelectionFilter& f) const { return !this->operator==(f); }
-
-    int filteredTypes() const;
-    bool isFiltered(SelectionFilterType type) const;
-    void setFiltered(SelectionFilterType type, bool filtered);
-
-    bool canSelect(const EngravingItem* element) const;
-    bool canSelectVoice(track_idx_t track) const;
-
-private:
-    unsigned int m_filteredTypes = static_cast<unsigned int>(SelectionFilterType::ALL);
 };
 
 //-------------------------------------------------------------------
