+++ src/notation/CMakeLists.txt
@@ -182,8 +182,9 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/view/internal/percussionnotepopupcontentmodel.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/internal/percussionnotepopupcontentmodel.h
 
-    ${CMAKE_CURRENT_LIST_DIR}/view/selectionfiltermodel.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/view/selectionfiltermodel.h
+    ${CMAKE_CURRENT_LIST_DIR}/view/selectionfilter/selectionfiltermodel.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/view/selectionfilter/selectionfiltermodel.h
+
     ${CMAKE_CURRENT_LIST_DIR}/view/editgridsizedialogmodel.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/editgridsizedialogmodel.h
